Item System, Inventory and small GUI elements
Item, Inventory and GUI elements were created, as of now the items are
only put into the inventory on start up within the Inventory class, GUI
is set up with a GUISkin, I have 3 items set up, it is very easy to set
up items. Code is documented enough to get started on creating Items,
inputting them into the database.